Extension: lists 11 #xj7
- Blocking retrieval with timeout

+++ src/parser.rs
@@ -290,7 +290,7 @@ impl Parser {
                             return Err(anyhow!("BLPOP command requires exactly two arguments"));
                         }
                         let key = self.extract_string(&elements[1])?;
-                        let timeout: u64 = self.extract_string(&elements[2])?.parse()?;
+                        let timeout: f64 = self.extract_string(&elements[2])?.parse()?;
                         Ok(RedisCommand::Blpop { key, timeout })
                     }
                     _ => Err(anyhow!("Unsupported command: {}", command_name)),
